# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TV Carlos API
  version: 3.0.0
  description: API pública e administrativa da Web TV. Escritas exigem cookie de sessão e token CSRF.
servers:
  - url: http://localhost:8082
    description: Docker local
...
security: []
paths:
...
  /api/auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Sessão atual, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/news:
    get:
      tags: [Content]
      operationId: listNews
      responses: { '200': { description: Notícias ordenadas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/News' } } } } } }
    post:
      tags: [Content]
      operationId: createNews
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/News' }
      responses: { '201': { description: Criada, content: { application/json: { schema: { $ref: '#/components/schemas/News' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/news/order:
    put:
      tags: [Content]
      operationId: reorderNews
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Order' }
      responses: { '204': { description: Reordenadas }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/news/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updateNews
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/News' }
      responses: { '200': { description: Atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/News' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Content]
      operationId: deleteNews
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Removida }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/grade:
    get:
      tags: [Content]
      operationId: listPrograms
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: search, in: query, schema: { type: string, maxLength: 100 } }
        - { name: category, in: query, schema: { type: string, maxLength: 100 } }
      responses: { '200': { description: Grade paginada e ordenada, content: { application/json: { schema: { $ref: '#/components/schemas/ProgramListResponse' } } } } }
    post:
      tags: [Content]
      operationId: createProgram
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Program' }
      responses: { '201': { description: Criado, content: { application/json: { schema: { $ref: '#/components/schemas/Program' } } } }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/grade/order:
    put:
      tags: [Content]
      operationId: reorderPrograms
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Order' }
      responses: { '204': { description: Reordenados }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/grade/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updateProgram
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Program' }
      responses: { '200': { description: Atualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Program' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Content]
      operationId: deleteProgram
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Removido }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/branding:
    get:
      tags: [Content]
      operationId: getBranding
      responses: { '200': { description: Identidade pública e dados legais, content: { application/json: { schema: { $ref: '#/components/schemas/Branding' } } } } }
    put:
      tags: [Content]
      operationId: updateBranding
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BrandingInput' } } } }
      responses: { '200': { description: Atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Branding' } } } }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/private-rooms:
    get:
      tags: [Content]
      operationId: listPrivateRooms
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Salas privadas cadastradas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PrivateRoom' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
    post:
      tags: [Content]
      operationId: createPrivateRoom
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/PrivateRoom' }
      responses: { '201': { description: Sala criada; a senha só é retornada nesta resposta, content: { application/json: { schema: { $ref: '#/components/schemas/CreatePrivateRoomResponse' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/private-rooms/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updatePrivateRoom
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/PrivateRoom' }
      responses: { '200': { description: Sala atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoom' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Content]
      operationId: deletePrivateRoom
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Sala removida e sessões privadas associadas revogadas }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-rooms/{id}/rotate-password:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Content]
      operationId: rotatePrivateRoomPassword
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '200': { description: Nova senha gerada; copie antes de sair da tela, content: { application/json: { schema: { $ref: '#/components/schemas/CreatePrivateRoomResponse' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-rooms/{id}/interaction:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Content]
      operationId: getPrivateRoomInteractionAdmin
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Configuração e mensagens da interação da sala, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInteractionAdmin' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-rooms/{id}/interaction/settings:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updatePrivateRoomInteractionSettings
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInteractionSettingsInput' } } } }
      responses: { '200': { description: Configuração de interação atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInteractionSettings' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-rooms/{id}/interaction/archive:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Content]
      operationId: archivePrivateRoomInteraction
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Mensagens da sala arquivadas }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-room-messages/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [Content]
      operationId: moderatePrivateRoomMessage
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomMessageModerationInput' } } } }
      responses: { '200': { description: Mensagem moderada, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomMessage' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/teacher/private-rooms:
    get:
      tags: [Content]
      operationId: listTeacherPrivateRooms
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Salas privadas visíveis para professor ou admin, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PrivateRoom' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/teacher/private-rooms/{id}/interaction:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Content]
      operationId: getTeacherPrivateRoomInteraction
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Interação da sala atribuída ao professor, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInteractionAdmin' } } } }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/teacher/private-rooms/{id}/material:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updateTeacherPrivateRoomMaterial
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomSupportMaterialInput' } } } }
      responses: { '200': { description: Material de apoio e página atual atualizados, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoom' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/private-room-access:
    post:
      tags: [Content]
      operationId: accessPrivateRoom
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomAccessInput' } } } }
      responses: { '200': { description: Acesso concedido; envia cookie HttpOnly tv_private_room_session, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomAccessResponse' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '429': { $ref: '#/components/responses/RateLimited' } }
  /api/private-room-access/{roomCode}:
    get:
      tags: [Content]
      operationId: getPrivateRoomAccess
      security: [{ privateRoomCookie: [] }]
      parameters: [{ name: roomCode, in: path, required: true, schema: { type: string, pattern: '^[a-zA-Z0-9-]{4,24}$' } }]
      responses: { '200': { description: Sala liberada para a sessão privada atual, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoom' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/private-room-access/{roomCode}/interaction:
    get:
      tags: [Content]
      operationId: getPrivateRoomInteractionPublic
      security: [{ privateRoomCookie: [] }]
      parameters: [{ name: roomCode, in: path, required: true, schema: { type: string, pattern: '^[a-zA-Z0-9-]{4,24}$' } }]
      responses: { '200': { description: Mensagens aprovadas e destaque público da sala, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInteractionPublic' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/private-room-access/{roomCode}/messages:
    post:
      tags: [Content]
      operationId: submitPrivateRoomMessage
      security: [{ privateRoomCookie: [] }]
      parameters: [{ name: roomCode, in: path, required: true, schema: { type: string, pattern: '^[a-zA-Z0-9-]{4,24}$' } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomMessageSubmitInput' } } } }
      responses: { '201': { description: Mensagem recebida para moderação ou publicação, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomMessageSubmitResponse' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '429': { $ref: '#/components/responses/RateLimited' } }
...
  /api/partners:
    get:
      tags: [Content]
      operationId: listPartners
      responses: { '200': { description: Parceiros ordenados, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Partner' } } } } } }
    post:
      tags: [Content]
      operationId: createPartner
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Partner' }
      responses: { '201': { description: Criado, content: { application/json: { schema: { $ref: '#/components/schemas/Partner' } } } }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/partners/order:
    put:
      tags: [Content]
      operationId: reorderPartners
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Order' }
      responses: { '204': { description: Reordenados }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/partners/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updatePartner
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Partner' }
      responses: { '200': { description: Atualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Partner' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Content]
      operationId: deletePartner
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Removido }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/header-links:
    get:
      tags: [Content]
      operationId: listHeaderLinks
      responses: { '200': { description: Até quatro botões ordenados, content: { application/json: { schema: { type: array, maxItems: 4, items: { $ref: '#/components/schemas/HeaderLink' } } } } } }
    post:
      tags: [Content]
      operationId: createHeaderLink
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/HeaderLink' }
      responses: { '201': { description: Criado, content: { application/json: { schema: { $ref: '#/components/schemas/HeaderLink' } } } }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/header-links/order:
    put:
      tags: [Content]
      operationId: reorderHeaderLinks
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/Order' }
      responses: { '204': { description: Reordenados }, '400': { $ref: '#/components/responses/ValidationError' } }
  /api/header-links/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Content]
      operationId: updateHeaderLink
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/HeaderLink' }
      responses: { '200': { description: Atualizado, content: { application/json: { schema: { $ref: '#/components/schemas/HeaderLink' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Content]
      operationId: deleteHeaderLink
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Removido }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/audit:
    get:
      tags: [Content]
      operationId: listAudit
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses: { '200': { description: Eventos recentes, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AuditLog' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: tv_session }
    privateRoomCookie: { type: apiKey, in: cookie, name: tv_private_room_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  requestBodies:
    News: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [text], properties: { text: { type: string, minLength: 1, maxLength: 500 } } } } } }
    Program: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProgramInput' } } } }
    PrivateRoom: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PrivateRoomInput' } } } }
...
    Order: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [ids], properties: { ids: { type: array, minItems: 1, maxItems: 500, uniqueItems: true, items: { type: string, format: uuid } } } } } } }
...
  responses:
    ValidationError: { description: Entrada rejeitada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Sessão ausente ou expirada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: Recurso inexistente, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Session: { type: object, required: [user, expiresAt], properties: { user: { type: object, required: [id, username, role], properties: { id: { type: string, format: uuid }, username: { type: string }, role: { type: string, enum: [admin, teacher] } } }, expiresAt: { type: string, format: date-time } } }
...
    News: { allOf: [{ $ref: '#/components/schemas/OrderedEntity' }, { type: object, required: [text], properties: { text: { type: string } } }] }
...
    Program: { allOf: [{ $ref: '#/components/schemas/OrderedEntity' }, { $ref: '#/components/schemas/ProgramInput' }] }
    ProgramListResponse: { type: object, required: [items, total], properties: { items: { type: array, items: { $ref: '#/components/schemas/Program' } }, total: { type: integer, minimum: 0 } } }
...
    PrivateRoom:
      allOf:
        - type: object
          required: [id, roomCode, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            roomCode: { type: string, pattern: '^[A-Z0-9-]{4,24}$' }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
        - $ref: '#/components/schemas/PrivateRoomInput'
...
    BrandingInput:
      type: object
      additionalProperties: false
      required: [companyName, tagline, watermarkText, logoText, logoUrl, scheduleTitle, tickerLabel, rssNewsUrl, partnerLabel, liveSource, liveYoutubeUrl, liveTitle, liveDescription, loopTitle, loopDescription, legalName, legalEmail, legalCnpj, legalCity, legalPhone]
      properties:
        companyName: { type: string, maxLength: 160 }
        tagline: { type: string, maxLength: 160 }
        watermarkText: { type: string, maxLength: 300 }
        logoText: { type: string, maxLength: 160 }
        logoUrl: { type: string, maxLength: 2048 }
        backgroundUrl: { type: string, maxLength: 2048, description: "Imagem de fundo do site; URL HTTPS, upload interno ou string vazia para usar o fundo padrão." }
        scheduleTitle: { type: string, maxLength: 160 }
        tickerLabel: { type: string, maxLength: 80 }
        rssNewsUrl: { type: string, maxLength: 2048, description: "URL HTTPS de feed RSS." }
        partnerLabel: { type: string, maxLength: 80 }
        liveSource: { type: string, enum: [obs, youtube], description: "Fonte do sinal principal ao vivo." }
        liveYoutubeUrl: { type: string, maxLength: 2048, description: "URL HTTPS da live do YouTube; use string vazia quando liveSource for obs." }
        liveTitle: { type: string, maxLength: 160 }
        liveDescription: { type: string, maxLength: 300 }
        loopTitle: { type: string, maxLength: 160 }
        loopDescription: { type: string, maxLength: 300 }
        legalName: { type: string, maxLength: 180 }
        legalEmail: { type: string, format: email }
        legalCnpj: { type: string, pattern: '^\d{2}\.\d{3}\.\d{3}/\d{4}-\d{2}$' }
        legalCity: { type: string, maxLength: 120 }
        legalPhone: { type: string, maxLength: 30 }
    Branding: { allOf: [{ $ref: '#/components/schemas/BrandingInput' }, { type: object, required: [updatedAt], properties: { updatedAt: { type: string, format: date-time } } }] }
...